feat: change report/[type] to report/post page
